# docs/swagger.yaml
openapi: 3.0.4
info:
  title: E-commerce API
  description: |
    Complete e-commerce API with user authentication, product management, 
    shopping cart, orders, payments, and admin functionality.
    
    ## Authentication
    - JWT Bearer token authentication
    - Email verification required for full access
    - Google OAuth support
    
    ## Roles
    - **user**: Regular customer permissions
    - **admin**: Full administrative access
    
    ## Rate Limiting
    - 1000 requests per hour per IP
    - 100 requests per minute for authenticated users
  version: 1.0.0
  termsOfService: https://example.com/terms
  contact:
    name: Abdulrahman Hegazy
    email: a.0hegazy@proton.me
...
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://ecommerce-express-boilerplate.onrender.com/api/v1
    description: Production server

# Global security - applies to all endpoints
security:
  - bearerAuth: []
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get user orders
      description: Retrieve paginated list of authenticated user's orders
      operationId: getOrders
      parameters:
        - $ref: '#/components/parameters/PaginationPage'
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Number of orders per page
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Orders list
                  value:
                    status: success
                    data:
                      orders:
                        - id: "64c9b8d3a7e9a2d8f0c1b238"
                          orderNumber: "ORD-2025-001"
                          userId: "64c9b8d3a7e9a2d8f0c1b235"
                          cartId: "64c9b8d3a7e9a2d8f0c1b234"
                          items:
                            - productId: "64c9b8d3a7e9a2d8f0c1b236"
                              quantity: 2
                              name: "Wireless Headphones"
                              price: 29.99
                              image: "https://example.com/headphones.jpg"
                              product:
                                id: "64c9b8d3a7e9a2d8f0c1b236"
                                name: "Wireless Headphones"
                                price: 29.99
                                image: "https://example.com/headphones.jpg"
                            - productId: "64c9b8d3a7e9a2d8f0c1b237"
                              quantity: 1
                              name: "Phone Case"
                              price: 15.99
                              image: "https://example.com/case.jpg"
                              product:
                                id: "64c9b8d3a7e9a2d8f0c1b237"
                                name: "Phone Case"
                                price: 15.99
                                image: "https://example.com/case.jpg"
                          subtotal: 75.97
                          tax: 6.08
                          shipping: 5.99
                          totalPrice: 88.04
                          shippingAddress:
                            street: "123 Main St"
                            city: "New York"
                            state: "NY"
                            zipCode: "10001"
                            country: "US"
                            phone: "+1234567890"
                          billingAddress:
                            street: "123 Main St"
                            city: "New York"
                            state: "NY"
                            zipCode: "10001"
                            country: "US"
                          status: "processing"
                          paymentStatus: "paid"
                          paymentMethod: "credit_card"
                          trackingNumber: "TRK123456789"
                          isActive: true
                          createdAt: "2025-10-15T14:00:00.000Z"
                          updatedAt: "2025-10-15T14:30:00.000Z"
                      pagination:
                        total: 5
                        currentPage: 1
                        totalPages: 1
                        pageSize: 10
                empty:
                  summary: No orders
                  value:
                    status: success
                    data:
                      orders: []
                      message: "You have no orders yet."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Orders
      summary: Create new order
      description: Create a new order from user's cart items
      operationId: makeOrder
      requestBody:
        required: true
        description: Order details and addresses
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              success:
                summary: Create order with addresses
                value:
                  shippingAddress:
                    street: "123 Main St"
                    city: "New York"
                    state: "NY"
                    zipCode: "10001"
                    country: "US"
                    phone: "+1234567890"
                  billingAddress:
                    street: "123 Main St"
                    city: "New York"
                    state: "NY"
                    zipCode: "10001"
                    country: "US"
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Order created
                  value:
                    status: success
                    data:
                      message: "Order placed successfully"
                      order:
                        id: "64c9b8d3a7e9a2d8f0c1b238"
                        orderNumber: "ORD-2025-001"
                        totalPrice: 88.04
                        status: "pending"
        '400':
          description: Bad request - Cart empty or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                emptyCart:
                  summary: Empty cart
                  value:
                    status: fail
                    data:
                      message: "Cart is empty or not found."
                insufficientStock:
                  summary: Insufficient stock
                  value:
                    status: fail
                    data:
                      message: "Insufficient stock for Wireless Headphones. Available: 5"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get single order
      description: Retrieve detailed information about a specific order
      operationId: getSingleOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: objectid
          description: Order ID
          example: "64c9b8d3a7e9a2d8f0c1b238"
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Order details
                  value:
                    status: success
                    data:
                      order:
                        id: "64c9b8d3a7e9a2d8f0c1b238"
                        orderNumber: "ORD-2025-001"
                        userId: "64c9b8d3a7e9a2d8f0c1b235"
                        cartId: "64c9b8d3a7e9a2d8f0c1b234"
                        items:
                          - productId: "64c9b8d3a7e9a2d8f0c1b236"
                            quantity: 2
                            name: "Wireless Headphones"
                            price: 29.99
                            image: "https://example.com/headphones.jpg"
                            product:
                              id: "64c9b8d3a7e9a2d8f0c1b236"
                              name: "Wireless Headphones"
                              price: 29.99
                              image: "https://example.com/headphones.jpg"
                          - productId: "64c9b8d3a7e9a2d8f0c1b237"
                            quantity: 1
                            name: "Phone Case"
                            price: 15.99
                            image: "https://example.com/case.jpg"
                            product:
                              id: "64c9b8d3a7e9a2d8f0c1b237"
                              name: "Phone Case"
                              price: 15.99
                              image: "https://example.com/case.jpg"
                        subtotal: 75.97
                        tax: 6.08
                        shipping: 5.99
                        totalPrice: 88.04
                        shippingAddress:
                          street: "123 Main St"
                          city: "New York"
                          state: "NY"
                          zipCode: "10001"
                          country: "US"
                          phone: "+1234567890"
                        billingAddress:
                          street: "123 Main St"
                          city: "New York"
                          state: "NY"
                          zipCode: "10001"
                          country: "US"
                        status: "processing"
                        paymentStatus: "paid"
                        paymentMethod: "credit_card"
                        paymentId: "pay_123456789"
                        paymentDate: "2025-10-15T14:05:00.000Z"
                        trackingNumber: "TRK123456789"
                        notes: "Handle with care"
                        isActive: true
                        createdAt: "2025-10-15T14:00:00.000Z"
                        updatedAt: "2025-10-15T14:30:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    patch:
      tags:
        - Orders
      summary: Cancel order
      description: Cancel a pending order (only if not processing/shipped/delivered and not paid)
      operationId: updateOrder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: objectid
          description: Order ID to cancel
          example: "64c9b8d3a7e9a2d8f0c1b238"
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Order cancelled
                  value:
                    status: success
                    data:
                      message: "Order cancelled successfully"
        '400':
          description: Bad request - Cannot cancel order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                cannotCancel:
                  summary: Cannot cancel
                  value:
                    status: fail
                    data:
                      message: "Order cannot be cancelled at this stage"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. For admin endpoints, user must have 'admin' role.
  parameters:
    PaginationPage:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Unauthorized. Please log in."
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Resource not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data: "Internal Server Error"
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, fail, error]
          example: success
        data:
          oneOf:
            - type: object
              description: Contains response data in success and fail cases
              properties:
                message:
                  type: string
                  example: "Successfully completed action"
                accessToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5..."
                user:
                  $ref: '#/components/schemas/User'
            - type: string
              description: Contains error message or additional info in error cases
              example: "Internal server error"
            - type: object
              nullable: true
              description: If no data is available, null can be provided
              example: null
      required:
        - status
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        data:
          type: string
          example: "Internal Server Error"
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "64c9b8d3a7e9a2d8f0c1b234"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        avatar:
          type: string
          format: uri
          example: "https://via.placeholder.com/150/cccccc/969696?text=User"
        phoneNumber:
          type: string
          example: "+1234567890"
        emailVerified:
          type: boolean
          example: true
        isActive:
          type: boolean
          example: true
        address:
          $ref: '#/components/schemas/Address'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: "2025-10-15T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-15T12:30:00.000Z"
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          default: "US"
          example: "US"
    UserPreferences:
      type: object
      properties:
        newsletter:
          type: boolean
          default: true
          example: true
        notifications:
          type: boolean
          default: true
          example: false
    CreateOrderRequest:
      type: object
      properties:
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        billingAddress:
          $ref: '#/components/schemas/BillingAddress'
    ShippingAddress:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          example: "US"
        phone:
          type: string
          example: "+1234567890"
    BillingAddress:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          example: "US"
...